+++ src/main/java/com/nbc/newsfeeds/domain/comment/code/CommentSuccessCode.java
@@ -1,4 +1,4 @@
-package com.nbc.newsfeeds.domain.comment;
+package com.nbc.newsfeeds.domain.comment.code;
 
 import org.springframework.http.HttpStatus;
